feat: ...

+++ src/pages/Clients/Clients.tsx
@@ -1,7 +1,8 @@
 import { useEffect, useState } from "react";
-import "./styles.scss";
 import { supabase } from "../../lib/supabase";
 import classNames from "classnames";
+import StatusIndicator from "../../components/StatusIndicator/StatusIndicator";
+import "./styles.scss";
 
 type Client = {
 	id: string;
@@ -19,11 +20,13 @@ const initClient = {
 
 export default function Clients() {
 	const [loading, setLoading] = useState(false);
-	const [error, setError] = useState(null);
+	const [error, setError] = useState<Error | string | null>(null);
 	const [clients, setClients] = useState<Client[]>([]);
 	const [client, setClient] = useState<Client>(initClient);
 	const [formVisible, setFormVisible] = useState(false);
+	const [deleteModalVisible, setDeleteModalVisible] = useState(false);
 	const [clientEditable, setClientEditable] = useState(false);
+	const [deleteInput, setDeleteInput] = useState("");
 
 	// TODO:
 	useEffect(() => {
@@ -76,20 +79,27 @@ export default function Clients() {
 		}
 	};
 
-	const deleteClient = async (clientId: string) => {
+	const deleteClient = async () => {
+		if (deleteInput !== "Alex please, don't break the program") {
+			setError("Please, enter correct sentence.");
+			return;
+		}
+
 		setLoading(true);
 		setError(null);
 		try {
 			const { error } = await supabase
 				.from("clients")
 				.delete()
-				.eq("id", clientId);
+				.eq("id", client.id);
 
 			if (error) throw error;
 
-			setClients((prev) => prev.filter((client) => client.id !== clientId));
+			setClients((prev) => prev.filter((c) => c.id !== client.id));
 
-			setFormVisible(false);
+			setClient(initClient);
+			setDeleteModalVisible(false);
+			setDeleteInput("");
 		} catch (error: any) {
 			setError(error.message);
 			console.error(error);
@@ -144,62 +154,60 @@ export default function Clients() {
 		}
 	};
 
-	console.log(error);
-	console.log(loading);
-
 	return (
 		<>
-			<div
+			<form
 				className={classNames("clients__form", {
 					"clients__form--visible": formVisible,
 				})}
+				onSubmit={(e) => {
+					e.preventDefault();
+					handleForm();
+				}}
 			>
-				<form
-					onSubmit={(e) => {
-						e.preventDefault();
-						handleForm();
-					}}
-				>
-					<div>
-						<label htmlFor="name">Name</label>
-						<input
-							onChange={(e) =>
-								handleFormOnChange(e.target.name, e.target.value)
-							}
-							value={client.name || ""}
-							name="name"
-							id="name"
-							type="text"
-							placeholder="Name"
-						/>
-					</div>
-					<div>
-						<label htmlFor="address">Address</label>
-						<input
-							onChange={(e) =>
-								handleFormOnChange(e.target.name, e.target.value)
-							}
-							value={client.address || ""}
-							name="address"
-							id="address"
-							type="text"
-							placeholder="Address"
-						/>
-					</div>
-					<div>
-						<label htmlFor="tel">Tel</label>
-						<input
-							onChange={(e) =>
-								handleFormOnChange(e.target.name, e.target.value)
-							}
-							value={client.tel || ""}
-							name="tel"
-							id="tel"
-							type="text"
-							placeholder="Tel"
-						/>
-					</div>
+				<p style={{ fontSize: "1.5rem" }}>
+					{clientEditable ? "Update client" : "Add new client"}
+				</p>
+				<div>
+					<label htmlFor="name">Name</label>
+					<input
+						className="clients__input"
+						onChange={(e) => handleFormOnChange(e.target.name, e.target.value)}
+						value={client.name || ""}
+						name="name"
+						id="name"
+						type="text"
+						placeholder="Name"
+						required
+					/>
+				</div>
+				<div>
+					<label htmlFor="address">Address</label>
+					<input
+						className="clients__input"
+						onChange={(e) => handleFormOnChange(e.target.name, e.target.value)}
+						value={client.address || ""}
+						name="address"
+						id="address"
+						type="text"
+						placeholder="Address"
+					/>
+				</div>
+				<div>
+					<label htmlFor="tel">Tel</label>
+					<input
+						className="clients__input"
+						onChange={(e) => handleFormOnChange(e.target.name, e.target.value)}
+						value={client.tel || ""}
+						name="tel"
+						id="tel"
+						type="text"
+						placeholder="Tel"
+					/>
+				</div>
+				<div className="clients__form-btn-container">
 					<button
+						className="clients__form-cancel-btn"
 						type="button"
 						onClick={() => {
 							setFormVisible(false);
@@ -209,10 +217,59 @@ export default function Clients() {
 					>
 						Cancel
 					</button>
-					<button type="submit">
-						{clientEditable ? "Update client" : "Add new client"}
+					<button className="clients__form-submit-btn" type="submit">
+						{loading
+							? clientEditable
+								? "Updating..."
+								: "Saving..."
+							: clientEditable
+								? "Update"
+								: "Save"}
 					</button>
-				</form>
+				</div>
+			</form>
+			<div
+				className={classNames("clients__delete-modal", {
+					"clients__delete-modal--visible": deleteModalVisible,
+				})}
+			>
+				<p style={{ fontSize: "1.5rem" }}>Delete client</p>
+				<p style={{ color: "var(--red-clr)" }}>
+					{error && (typeof error === "string" ? error : error.message)}
+				</p>
+				<p>
+					Are you absolutely sure you want to delete this client{" "}
+					<span style={{ fontWeight: "500" }}>{client.name}</span>?
+				</p>
+				<p>
+					Type in the sentence "
+					<span style={{ color: "var(--primary-clr)" }}>
+						Alex please, don't break the program
+					</span>
+					" below.
+				</p>
+				<input
+					className="clients__input"
+					onChange={(e) => setDeleteInput(e.target.value)}
+					type="text"
+					value={deleteInput}
+					required
+				/>
+				<div>
+					<button
+						onClick={() => {
+							setDeleteModalVisible(false);
+							setClient(initClient);
+							setDeleteInput("");
+						}}
+						disabled={loading}
+					>
+						Cancel
+					</button>
+					<button disabled={loading} onClick={deleteClient}>
+						{loading ? "Deleting..." : "Delete"}
+					</button>
+				</div>
 			</div>
 			<div
 				onClick={() => {
@@ -221,7 +278,7 @@ export default function Clients() {
 					setClient(initClient);
 				}}
 				className={classNames("curtain", {
-					"curtain--visible": formVisible,
+					"curtain--visible": formVisible || deleteModalVisible,
 				})}
 			></div>
 			<section className="clients__section">
@@ -232,13 +289,9 @@ export default function Clients() {
 				{clients.map((client) => {
 					return (
 						<div className="clients__client-card" key={client.id}>
-							<div>
-								<p>Name: {client.name ? client.name : "Unknown"}</p>
-								<p>Address: {client.address ? client.address : "Unknown"}</p>
-								<p>Tel: {client.tel ? client.tel : "Unknown"}</p>
-							</div>
-							<div>
+							<div className="clients__btn-container">
 								<button
+									className="clients__edit-btn"
 									onClick={() => {
 										setFormVisible(true);
 										setClient(client);
@@ -247,12 +300,30 @@ export default function Clients() {
 								>
 									Edit
 								</button>
-								<button onClick={() => deleteClient(client.id)}>Delete</button>
+								<button
+									className="clients__delete-btn"
+									// onClick={() => deleteClient(client.id)}
+									onClick={() => {
+										setClient(client);
+										setDeleteModalVisible(true);
+									}}
+								>
+									Delete
+								</button>
+							</div>
+							<div>
+								<p className="clients__label">Name</p>
+								<p className="clients__details">{client.name ?? ""}</p>
+								<p className="clients__label">Address</p>
+								<p className="clients__details">{client.address ?? ""}</p>
+								<p className="clients__label">Tel</p>
+								<p className="clients__details">{client.tel ?? ""}</p>
 							</div>
 						</div>
 					);
 				})}
 			</section>
+			<StatusIndicator loading={loading} error={error} />
 		</>
 	);
 }
